requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
plotly>=5.15.0
python-dateutil>=2.8.0
PyGithub>=2.1.0
requests>=2.31.0
cryptography>=41.0.0
gspread>=5.10.0
google-auth>=2.22.0